// src/types/others.ts
export interface SearchObj {
    page: number;
    limit: number;
    order: string;
}

export interface ProductSearchObj {
    page: number;
    limit: number;
    order: string;
    restaurant_mb_id?: string,
    product_collection?: string
}

export interface MemberLiken {
    like_group: string,
    like_status: number,
    like_ref_id: string
}